requirements.txt
Pillow>=5.1
requests>=2.12.5
emoji>=0.4.5